array storage 추가
